New agggregate table.
To migrate:

* flash shell
* from database import Aggregate
* Aggregate.migrate()

A single entry for each device ID is stored in the 'aggregate' table,
which is used for page generation. This limits the amount of enteries
that need to be run through mongo's aggregation framework from
(number of devices) * (number of days of data stored) to (number of
devices).

Change-Id: I40cb68786b44bf4afa2e173f4027eabe2e861564